// resources/scripts/interfaces/SupplyCharacteristicValueInterface.ts
export interface SupplyCharacteristicValueInterface {
  supplieCaracteristicValueId: number | null
  supplieCaracteristicId: number
  supplieId: number
  supplieCaracteristicValueValue: string
  supplieCaracteristicValueCreatedAt: Date | null
  supplieCaracteristicValueUpdatedAt: Date | null
  deletedAt: Date | null
}